Modularize helper functions
