+++ src/components/Navigation.js
@@ -1,38 +1,10 @@
 import { useState } from 'react';
 import { useSelector } from 'react-redux';
-import { NavLink } from 'react-router-dom';
+import { generateNavLinks, searchByRank } from './functions';
 
 const Navigation = () => {
   const { cryptoArray } = useSelector((store) => store.crypto);
   const [querry, setQuerry] = useState('');
-
-  const generateNavLinks = (cryptoArray) => {
-    const linkArray = [];
-    for (let rank = 1; rank <= cryptoArray.length; rank += 1) {
-      linkArray.push(
-        <NavLink key={rank} className="link" to={`/${rank}`}>
-          {`#${cryptoArray[rank - 1].rank} ${cryptoArray[rank - 1].name}`}
-        </NavLink>,
-      );
-    }
-    return linkArray;
-  };
-
-  const searchByRank = (rank) => {
-    const token = cryptoArray.filter((coin) => coin.rank === rank);
-
-    if (token.length === 0) {
-      return <p>Out Of Bounds.</p>;
-    }
-
-    const result = (
-      <NavLink key={rank} className="link" to={`/${token[0].rank}`}>
-        {`#${token[0].rank} ${token[0].name}`}
-      </NavLink>
-    );
-
-    return result;
-  };
 
   return (
     <nav>
@@ -46,15 +18,16 @@ const Navigation = () => {
       {querry
         ? (
           <div>
-            {searchByRank(querry)}
+            {searchByRank(cryptoArray, querry)}
           </div>
         )
         : (
-          <ul className="column">
+          <ul>
             {generateNavLinks(cryptoArray)}
           </ul>
         )}
     </nav>
   );
 };
+
 export default Navigation;
